requirements.txt
Django==1.9.7
psycopg2==2.6.1
django-solo==1.1.2
django-geoposition==0.2.3
